Update parser
- Automatically upgrade BEL builtin molecular activities
- Reorganize exceptions
- Fix namespace summary and related tests,  now includes namespaces in edges
- Reorganize parser output structure (remove TARGET constant)
- Update all relevant tests
- Automatically add GO namespace to concept parser
