+++ docs/source/reference/logging.rst
@@ -1,11 +1,4 @@
 Logging Messages
 ================
-Errors
-------
 .. automodule:: pybel.exceptions
-    :members:
-
-Parse Exceptions
-----------------
-.. automodule:: pybel.parser.exc
     :members:
