+++ src/pybel/io/line_utils.py
@@ -13,12 +13,12 @@ from tqdm import tqdm
 
 from bel_resources import ResourceError, split_file_to_annotations_and_definitions
 from ..constants import INVERSE_DOCUMENT_KEYS, REQUIRED_METADATA
-from ..manager import Manager
-from ..parser import BELParser, MetadataParser
-from ..parser.exc import (
+from ..exceptions import (
     BELParserWarning, BELSyntaxError, InconsistentDefinitionError, MalformedMetadataException, MissingMetadataException,
     PlaceholderAminoAcidWarning, VersionFormatWarning,
 )
+from ..manager import Manager
+from ..parser import BELParser, MetadataParser
 from ..struct.graph import BELGraph
 
 __all__ = [
